test: split slow tests
